Allow TLS options to be set and passed to https.request()
The node-fetch Request object uses the platform https.request() API
to send the request. This API allows for additional TLS options to
be set to control how the connection is established. This can be
necessary when an abstraction layer sets the agent field to route
the request through a forward proxy, i.e., the consumer is unable
to use the agent to set TLS options (see
https://github.com/node-fetch/node-fetch/issues/904).